Created custom color set for Fonts
Created a custom color set for fonts, to be used in tandem with the custom color sets created as the background color of the list items tags. This will ensure that the font color toggles between black and white, based on whether the app is in light or dark mode.